fix: block deleting hardware that is currently in use, add demo reset script

+++ backend/src/main/java/com/hardwarehub/service/HardwareService.java
@@ -76,7 +76,11 @@ public class HardwareService {
     }
 
     public void delete(Long hardwareId) {
-        getOrThrow(hardwareId);
+        Hardware hardware = getOrThrow(hardwareId);
+        if (hardware.getStatus() == HardwareStatus.IN_USE) {
+            throw new IllegalHardwareStateException(
+                    "cannot delete hardware " + hardwareId + ": it is currently in use, return it first");
+        }
         hardwareRepository.deleteById(hardwareId);
     }
